Automatic resizing
Using the os library I created 2 global variables that save the size of the terminal when program starts.
Then in a while loop the program checks if the size of the terminal has changed and if it did it updates the video and subtitle windows.
Also disabled the cursor so it doesn't ruin the experience.

No new dependency, only the os library was added.
I have changed the import of cv2 library.

Drawbacks: When the subtitle window is resized the current subtitles might disappear, but only until new subtitles are rendered. I think this is caused by subtitle_show function and how it's implemented.